# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SETU-DRR API",
    "description": "Hazard Red Zone Identification, Carrying Capacity Assessment & Relocation Decision Support Platform.",
    "version": "v1.0.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication & Identity"
        ],
        "summary": "Authenticate user with email and password",
        "description": "Verifies credentials using Argon2id, creates a secure server-side session, and sets an HTTP-only session cookie. Returns the safe authenticated identity. Protected by configurable sliding-window rate limiting.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid parameters or malformed input format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated - Missing, invalid, expired, or revoked session cookie.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit exceeded. Please try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication & Identity"
        ],
        "summary": "Get authenticated user identity",
        "description": "Resolves the current authenticated user identity from the session cookie. Returns 401 if unauthenticated, expired, or revoked.",
        "operationId": "get_me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated - Missing, invalid, expired, or revoked session cookie.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication & Identity"
        ],
        "summary": "Revoke session and log out",
        "description": "Revokes the active server-side session and clears the session cookie.",
        "operationId": "logout_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication & Identity"
        ],
        "summary": "Register a new civilian user",
        "description": "Public self-registration for civilian users. Privileged roles (GOVERNMENT_OFFICIAL, SYSTEM_ADMIN) cannot be selected and are strictly rejected.",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid parameters or malformed input format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Standardized machine-readable error code string."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation of the error."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Unique request tracing identifier."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "description": "Contextual error details or validation breakdown."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Standardized error details conforming to SETU-DRR API error envelope."
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "Error details payload."
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Canonical API error response envelope."
      },
...
      "JurisdictionDTO": {
        "properties": {
          "admin_id": {
            "type": "integer",
            "title": "Admin Id",
            "description": "Canonical administrative boundary ID (admin_boundary.id)."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Administrative boundary name (e.g. Wayanad, Kodagu)."
          },
          "level": {
            "type": "string",
            "title": "Level",
            "description": "Administrative boundary level (e.g. district, state)."
          },
          "lgd_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lgd Code",
            "description": "Local Government Directory (LGD) code."
          }
        },
        "type": "object",
        "required": [
          "admin_id",
          "name",
          "level"
        ],
        "title": "JurisdictionDTO",
        "description": "Authoritative administrative jurisdiction assignment."
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email",
            "description": "User account email address.",
            "examples": [
              "officer@setu.gov.in"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password",
            "description": "Account password.",
            "examples": [
              "********"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Request payload for user authentication."
      },
      "LogoutResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Status message.",
            "default": "Logged out successfully."
          }
        },
        "type": "object",
        "title": "LogoutResponse",
        "description": "Response returned upon successful session revocation."
      },
...
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email",
            "description": "Valid email address.",
            "examples": [
              "citizen@example.org"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "Secure password (minimum 8 characters)."
          },
          "full_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name",
            "description": "Full name of the user.",
            "examples": [
              "Asha Nair"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password",
          "full_name"
        ],
        "title": "RegisterRequest",
        "description": "Request payload for public civilian registration.\n\nRole is strictly not accepted from client and is automatically set to CIVILIAN."
      },
      "Role": {
        "type": "string",
        "enum": [
          "CIVILIAN",
          "GOVERNMENT_OFFICIAL",
          "SYSTEM_ADMIN"
        ],
        "title": "Role",
        "description": "User identity roles for SETU-DRR authentication."
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique user identifier."
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "User email address."
          },
          "full_name": {
            "type": "string",
            "title": "Full Name",
            "description": "Full name of user."
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "User role (CIVILIAN, GOVERNMENT_OFFICIAL, SYSTEM_ADMIN)."
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether user account is active."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Account creation timestamp."
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At",
            "description": "Timestamp of most recent successful login."
          },
          "jurisdiction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JurisdictionDTO"
              },
              {
                "type": "null"
              }
            ],
            "description": "Administrative jurisdiction assigned to privileged user, or None for unconstrained/civilian users."
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "Safe authenticated identity DTO.\n\nGuaranteed never to expose password hashes, session tokens, or internal secrets."
      },
...
    }
  }
}